requirements.txt
Flask
python-dotenv
waitress
requests
psycopg2-binary
